Cleanup ipc implementation
