+++ src/vs/platform/webview/electron-main/webviewIpcs.ts
@@ -3,11 +3,15 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { IServerChannel } from 'vs/base/parts/ipc/common/ipc';
 import { Event } from 'vs/base/common/event';
-import { webContents } from 'electron';
+import { IServerChannel } from 'vs/base/parts/ipc/common/ipc';
+import { IWebviewMainService } from 'vs/platform/webview/common/webviewMainService';
 
 export class WebviewChannel implements IServerChannel {
+
+	constructor(
+		@IWebviewMainService private readonly webviewMainService: IWebviewMainService,
+	) { }
 
 	listen(_: unknown, event: string): Event<any> {
 		throw new Error(`Event not found: ${event}`);
@@ -15,19 +19,9 @@ export class WebviewChannel implements IServerChannel {
 
 	async call(_: unknown, command: string, arg?: any): Promise<any> {
 		switch (command) {
-			case 'setIgnoreMenuShortcuts': this.setIgnoreMenuShortcuts(arg[0], arg[1]); return;
+			case 'setIgnoreMenuShortcuts': this.webviewMainService.setIgnoreMenuShortcuts(arg[0], arg[1]); return;
 		}
 
 		throw new Error(`Call not found: ${command}`);
-	}
-
-	private setIgnoreMenuShortcuts(webContentsId: number, enabled: boolean) {
-		const contents = webContents.fromId(webContentsId);
-		if (!contents) {
-			throw new Error(`Invalid webContentsId: ${webContentsId}`);
-		}
-		if (!contents.isDestroyed()) {
-			contents.setIgnoreMenuShortcuts(enabled);
-		}
 	}
 }
